authorization for categories Authorization for categories added so that only an admin can view, create, edit or delete a category. Users get a 404 page if they attempt to view admin namespaced urls.
