+++ app/views/admin/categories/edit.html.erb
@@ -1,4 +1,4 @@
-<%= form_for(@category) do |f| %>
+<%= form_for([:admin, @category]) do |f| %>
   <%= f.label :title %>
   <%= f.text_field :title %><br>
   <%= f.submit %>
