+++ app/views/admin/categories/index.html.erb
@@ -5,3 +5,5 @@
   <li><%= link_to category.title, admin_category_path(category) %></li>
   <% end %>
 </ul>
+
+<%= link_to "Create New Category", new_admin_category_path%>
