+++ spec/features/admin_can_delete_an_exisiting_category_spec.rb
@@ -2,7 +2,8 @@ require 'rails_helper'
 
 RSpec.feature "Deleting a Category", type: :feature do
     scenario "user can delete an existing category" do
-      # admin = login_admin
+      admin = User.create(first_name: "Lucy", last_name: "Fox", username: "lucyfox", password: "password", role: 1)
+      allow_any_instance_of(ApplicationController).to receive(:current_user).and_return(admin)
 
       c = Category.create(title: "Health")
       c1 = Category.create(title: "Nutrition")
