+++ spec/features/admin_can_edit_existing_category_spec.rb
@@ -2,7 +2,8 @@ require 'rails_helper'
 
 RSpec.feature "Edit a Category", type: :feature do
     scenario "user can edit a category" do
-      # admin = login_admin
+      admin = User.create(first_name: "Lucy", last_name: "Fox", username: "lucyfox", password: "password", role: 1)
+      allow_any_instance_of(ApplicationController).to receive(:current_user).and_return(admin)
 
       c = Category.create(title: "Nutrition")
