--- spec/features/user_can_view_specific_category_spec.rb
@@ -1,11 +1,0 @@
-require 'rails_helper'
-
-RSpec.feature "Viewing a Category", type: :feature do
-    scenario "user can view an existing category" do
-      category = Category.create(title: "Nutrition")
-
-      visit admin_category_path(category)
-
-      expect(page).to have_content("Nutrition")
-    end
-end
